+++ test_imports.py
@@ -3,7 +3,7 @@ Test script to verify imports are working correctly.
 """
 from rpg_game.character import Character, Boss
 from rpg_game.weapon import Weapon
-from rpg_game.utils.logger import GameLogger
+from rpg_game.game_logger import GameLogger
 from rpg_game.utils.console import print_border